optimized world

+++ include/stellarlib/ecs/bitset.hpp
@@ -26,11 +26,11 @@
 
 #include <stellarlib/ext/memory.hpp>
 
-#include <cstddef>
+#include <cstdint>
 
 namespace stellarlib::ecs::internal
 {
-class bitset final : ext::vector_allocator<std::size_t>
+class bitset final : ext::vector_allocator<std::uintmax_t>
 {
 public:
 	[[nodiscard]]
@@ -50,10 +50,10 @@ public:
 
 	~bitset() noexcept;
 
-	void insert(std::size_t bit) noexcept;
+	void insert(std::uintmax_t bit) noexcept;
 
 	[[nodiscard]]
-	auto contains(std::size_t bit) const noexcept
+	auto contains(std::uintmax_t bit) const noexcept
 		-> bool;
 
 	[[nodiscard]]
@@ -68,15 +68,15 @@ public:
 	auto operator>=(const bitset &other) const noexcept
 		-> bool;
 
-	void erase(std::size_t bit) noexcept;
+	void erase(std::uintmax_t bit) noexcept;
 
 	void clear() noexcept;
 
 private:
-	std::size_t _size{};
-	std::size_t _capacity{};
-	std::size_t *_begin{};
-	std::size_t *_end{};
+	std::uintmax_t _size{};
+	std::uintmax_t _capacity{};
+	std::uintmax_t *_begin{};
+	std::uintmax_t *_end{};
 };
 }
